Fix wrong date/games when activity remains alive during day changes
This fixes the issue where the games screen will show an incorrect date and games when the app is re-opened after a day.
For example, if the app is launched on Jan/01 the first page in the games screen will show "Today" and the games that are played on Jan/01. Then, if the application is stopped (by navigating away from it for example) and is re-opened on Jan/02, the games screen will show "Today" with the games for Jan/01, which not what the user expects to see.

After this fix, upon re-opening the app on Jan/02 the games screen will show "Today" AND the games that are played on Jan/02.